Add e2e tests for standalone running flows
- Run on cluster
- Run as a workflow
